move navbar components
