+++ apps/nextjs/src/app/(main)/@navbar/default.tsx
@@ -1,4 +1,4 @@
-import Navbar from "~/components/navbar/navbar";
+import Navbar from "~/app/(main)/@navbar/_navbar-components/navbar";
 
 export default function NavbarDefaultSlot() {
   return <Navbar />;
